==> 05-computcorr.Rmd <==
# Computing correlations among genes {#computcorr}
In this section, correlation matrices among genes are computed from the
normalized read count matrices, which will be used for downstream analyses.

Load the data.
```{r, eval = FALSE}
day1_norm <- readRDS(file = "backup/01_005_day1_norm_normalized.rds")
day7_hypo <- readRDS(file = "backup/02_005_day7_hypo_normalized.rds")
sc68_vehi <- readRDS(file = "backup/03_005_sc68_vehi_normalized.rds")
sc68_cisp <- readRDS(file = "backup/04_005_sc68_cisp_normalized.rds")
pbmc_4000 <- readRDS(file = "backup/05_005_pbmc_4000_normalized.rds")
pbmc_6000 <- readRDS(file = "backup/06_005_pbmc_6000_normalized.rds")
```

The following function `do_cor_variables()` computes a correlation matrix from
the input data (ASURAT object).
Users can choose a measure of correlation coefficient by setting `method`.
Multiple methods are also accepted, e.g., `method = "spearman"`,
and `method = c("pearson", "spearman", "kendall")`.
```{r, eval = FALSE}
do_cor_variables <- function(obj, method){
  res <- list()
  tmat <- t(obj[["data"]][["log1p"]])
  for(m in method){
    res <- c(res, list(cor(tmat, method = m)))
  }
  names(res) <- method
  return(res)
}

day1_norm_cor <- do_cor_variables(obj = day1_norm, method = c("spearman"))
day7_hypo_cor <- do_cor_variables(obj = day7_hypo, method = c("spearman"))
sc68_vehi_cor <- do_cor_variables(obj = sc68_vehi, method = c("spearman"))
sc68_cisp_cor <- do_cor_variables(obj = sc68_cisp, method = c("spearman"))
pbmc_4000_cor <- do_cor_variables(obj = pbmc_4000, method = c("spearman"))
pbmc_6000_cor <- do_cor_variables(obj = pbmc_6000, method = c("spearman"))
```

Save the objects.
```{r, eval = FALSE}
saveRDS(day1_norm_cor, file = "backup/01_006_day1_norm_correlation.rds")
saveRDS(day7_hypo_cor, file = "backup/02_006_day7_hypo_correlation.rds")
saveRDS(sc68_vehi_cor, file = "backup/03_006_sc68_vehi_correlation.rds")
saveRDS(sc68_cisp_cor, file = "backup/04_006_sc68_cisp_correlation.rds")
saveRDS(pbmc_4000_cor, file = "backup/05_006_pbmc_4000_correlation.rds")
saveRDS(pbmc_6000_cor, file = "backup/06_006_pbmc_6000_correlation.rds")
```
